add bbl config for external topgun-worker
this is an external worker configured to register with the prod concourse but
also have connectivity to its own dedicated bosh director for running topgun.

the create-topgun.sh script doesn't actually create anything, but it's a
placeholder that you can run to at least make sure everything's wired properly
and get the creds you need to mess around with stuff.

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>
Co-authored-by: James Thomson <jthomson@pivotal.io>